Improve deps
