+++ packages/schema-wrapping/package.json
@@ -25,9 +25,6 @@
     "@graphql-tools/utils": "5.0.0",
     "tslib": "1.11.1"
   },
-  "devDependencies": {
-    "@graphql-tools/mocking": "5.0.0"
-  },
   "publishConfig": {
     "access": "public",
     "directory": "dist"
