+++ packages/schema-wrapping/src/wrap/generateProxyingResolvers.ts
@@ -1,10 +1,7 @@
 import { GraphQLSchema, GraphQLFieldResolver, GraphQLObjectType } from 'graphql';
 
-import { Transform, Operation, SubschemaConfig, getResponseKeyFromInfo, getErrors } from '@graphql-tools/utils';
-import { delegateToSchema } from '../delegate/delegateToSchema';
-import { handleResult } from '../delegate/results/handleResult';
-
-import { getSubschema } from '../delegate/subschema';
+import { Transform, Operation, getResponseKeyFromInfo, getErrors } from '@graphql-tools/utils';
+import { delegateToSchema, getSubschema, handleResult, SubschemaConfig } from '../delegate';
 
 export function generateProxyingResolvers({
   subschemaConfig,
