+++ packages/schema-wrapping/src/wrap/index.ts
@@ -4,3 +4,4 @@ export { defaultCreateProxyingResolver } from './generateProxyingResolvers';
 export * from './transforms/index';
 
 export { makeRemoteExecutableSchema, defaultCreateRemoteResolver } from './makeRemoteExecutableSchema';
+export * from './types';
