+++ packages/schema-wrapping/src/wrap/types.ts
@@ -1,0 +1,10 @@
+import { GraphQLSchema, GraphQLFieldResolver, BuildSchemaOptions } from 'graphql';
+import { Executor, Subscriber } from '../delegate';
+
+export interface IMakeRemoteExecutableSchemaOptions {
+  schema: GraphQLSchema | string;
+  executor?: Executor;
+  subscriber?: Subscriber;
+  createResolver?: (executor: Executor, subscriber: Subscriber) => GraphQLFieldResolver<any, any>;
+  buildSchemaOptions?: BuildSchemaOptions;
+}
