+++ packages/utils/package.json
@@ -19,8 +19,6 @@
     "graphql": "^14.0.0 || ^15.0.0"
   },
   "devDependencies": {
-    "@graphql-tools/schema-generator": "5.0.0",
-    "@graphql-tools/schema-wrapping": "5.0.0",
     "@types/dateformat": "3.0.1",
     "dateformat": "3.0.3",
     "graphql-scalars": "1.1.2"
